Make long

+++ src/main/java/com/nuclearw/rapsheet/Record.java
@@ -38,7 +38,7 @@ public class Record {
 
 	@NotNull
 	@Column(name="time")
-	private int time;
+	private long time;
 
 	@NotNull
 	@Column(name="sealed")
@@ -96,11 +96,11 @@ public class Record {
 		this.chargeLong = chargeLong;
 	}
 
-	public int getTime() {
+	public long getTime() {
 		return time;
 	}
 
-	public void setTime(int time) {
+	public void setTime(long time) {
 		this.time = time;
 	}
